create: repositorios de cada entidad, add: notacion jpa a las entidades

+++ backend/src/main/java/ar/edu/utn/dsi/ppai/repositories/AlcanceSismoRepository.java
@@ -6,6 +6,6 @@ import ar.edu.utn.dsi.ppai.entities.AlcanceSismo;
 
 // Repositorio de prueba
 @Repository
-public interface AlcanceSismoRepository extends JpaRepository<AlcanceSismo, Integer> {
+public interface AlcanceSismoRepository extends JpaRepository<AlcanceSismo, Long> {
     // Sin metodos, solo hereda los de spring boot
 }
